+++ src/cli/cli.js
@@ -1,5 +1,6 @@
 import install from './commands/install.js'
 import add from './commands/add.js'
+import remove from './commands/remove.js'
 
 import debug from './options/debug.js'
 import use from './options/use.js'
@@ -9,7 +10,8 @@ import info from './options/info.js'
 
 const commands = [
   install,
-  add
+  add,
+  remove
 ]
 
 const options = {