+++ src/cli/commands/remove.js
@@ -1,10 +1,10 @@
 import { translateFlag } from '../../helpers/args.js'
 
-const add = {
-  command: 'add <package> [FLAGS]',
-  aliases: ['a'],
-  desc: 'add package',
-  conflicts: ['install, remove'],
+const remove = {
+  command: 'remove <package>',
+  aliases: ['r, rm, uninstall, un'],
+  desc: 'remove package',
+  conflicts: ['install, add'],
   builder: (yargs) => {
     yargs.positional('package', {
       type: 'string',
@@ -14,8 +14,8 @@ const add = {
     yargs.option('save-dev', {
       alias: 'D',
       type: 'boolean',
-      desc: 'add package as devDependencies',
-      usage: '$0 add <package> --save-dev',
+      desc: 'remove package as devDependencies',
+      usage: '$0 remove <package> --save-dev',
       implies: ['package'],
       conflicts: ['save-optional, save-peer']
     })
@@ -23,33 +23,25 @@ const add = {
     yargs.option('save-optional', {
       alias: 'O',
       type: 'boolean',
-      desc: 'add package as optionalDependencies',
-      usage: '$0 add <package> --save-optional',
+      desc: 'remove package as optionalDependencies',
+      usage: '$0 remove <package> --save-optional',
       implies: ['package'],
       conflicts: ['save-dev, save-peer']
     })
 
     yargs.option('save-peer', {
       type: 'boolean',
-      desc: 'add package as peerDependencies',
-      usage: '$0 add <package> --save-peer',
+      desc: 'remove package as peerDependencies',
+      usage: '$0 remove <package> --save-peer',
       implies: ['package'],
       conflicts: ['save-dev', 'save-optional']
-    })
-
-    yargs.option('save-exact', {
-      alias: 'E',
-      type: 'boolean',
-      desc: 'add package as devDependencies',
-      usage: '$0 add <package> --save-exact',
-      implies: ['package']
     })
 
     yargs.option('global', {
       alias: 'g',
       type: 'boolean',
-      desc: 'add package as global',
-      usage: '$0 add <package> --global',
+      desc: 'remove package as global',
+      usage: '$0 remove <package> --global',
       implies: ['package']
     })
 
@@ -65,14 +57,10 @@ const add = {
       if ('save-peer' in yargs) {
         translateFlag('--save-peer')
       }
-
-      if ('save-exact' in yargs) {
-        translateFlag('--save-exact', '-E')
-      }
     })
 
     return yargs
   }
 }
 
-export default add
+export default remove