+++ src/cli/middleware.js
@@ -1,5 +1,6 @@
 import { argv } from 'node:process'
 import { cleanFlag, translateFlag } from '../helpers/args.js'
+import { translateCommand } from '../helpers/cmds.js'
 import { detectVoltaPin, getCurrentPackageManager } from '../helpers/get.js'
 import { getPackageConfiguration } from '../packages/list.js'
 
@@ -40,11 +41,15 @@ const middleware = async (yargs) => {
   }
 
   if (globalThis?.pkg?.cmd) {
-    globalThis.pkg.config = await getPackageConfiguration(globalThis.pkg.cmd)
+    globalThis.pkg.config = await getPackageConfiguration()
   }
 
   if ('global' in yargs) {
     translateFlag('--global', '-g')
+  }
+
+  if (yargs._.length) {
+    translateCommand(yargs._[0])
   }
 }