+++ src/packages/managers/yarn.js
@@ -5,12 +5,18 @@ const yarn = {
   version: '',
   lockFile: 'yarn.lock',
   cmds: {
-    i: 'install'
+    i: 'install',
+    a: 'add',
+    r: 'remove',
+    rm: 'remove',
+    un: 'remove'
   },
   args: {
     '--global': ['global', 1],
     '-g': ['global', 1],
     '--save-dev': '--dev',
+    '--save-optional': '--optional',
+    '--save-peer': '--peer',
     '--save-exact': '--exact',
     '-L': '--latest'
   }